requirements.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-filter==2.0.0
django-redis==4.9.0
django-redis-sessions==0.6.1
djangorestframework==3.8.2
idna==2.7
kombu==4.2.2.post1
Markdown==2.6.11
mysqlclient==1.3.13
oauthlib==2.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
sqlparse==0.2.4
urllib3==1.23
uWSGI==2.0.17.1
vine==1.2.0
